feat: add UML diagrams and refractor data types

+++ src/main/kotlin/models/Cercle.kt
@@ -1,23 +1,24 @@
 package models
+
 import utils.*
 
 /**
  * Definició de la classe Cercle que hereta de la classe abstracte ObjecteGeometric
  * @author itec-bcn
  */
-public class Cercle: ObjecteGeometric{
+public class Cercle : ObjecteGeometric {
     private var radi: Double
 
     /**
      * Definició del constructor de la classe Cercle que crida al constructor de la classe pare
      * per a inicialitzar els atributs del pare
      *
-     * @param x de tipus Int per especificar la posició de la figura sobre l'eix de horitzontal
-     * @param y de tipus Int per a especificar la posició de la figura sobre l'eix vertical
+     * @param x de tipus Double per especificar la posició de la figura sobre l'eix de horitzontal
+     * @param y de tipus Double per a especificar la posició de la figura sobre l'eix vertical
      * @param color de tipus String per a especificar el color de la figura
      * @param radi de tipus Double per a especificar el radi de la figura
      */
-    constructor(x:Int, y:Int, color:String, radi: Double) : super(x, y, color) {
+    constructor(x: Double, y: Double, color: String, radi: Double) : super(x, y, color) {
         this.radi = radi
     }
 
@@ -43,7 +44,7 @@ public class Cercle: ObjecteGeometric{
      * Mètode getter que retorna el radi del Cercle
      * @return this.radi Double
      */
-    public fun getRadi(): Double{
+    public fun getRadi(): Double {
         return this.radi
     }
